requirements.txt
bidict
blinker
cachetools
click
colorama
Flask
Flask-Cors
Flask-SocketIO
google-auth
google-cloud
google-cloud-vision
itsdangerous
Jinja2
MarkupSafe
pyasn1
pyasn1-modules
python-engineio
python-socketio
rsa
six
urllib3==1.26.6
Werkzeug
# eventlet==0.32.0
# gunicorn==20.1.0
# dnspython
gunicorn==20.1.0 
eventlet==0.30.2